% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertSexFormat.R
\name{ConvertSexFormat}
\alias{ConvertSexFormat}
\title{Convert Sex Data to format used by our Functions}
\usage{
ConvertSexFormat(Sex, female.string = "F", male.string = "M")
}
\arguments{
\item{Sex}{A character or factor vector (vector produced will be the same type) that includes the sex of each individual.}

\item{female.string}{How sex for females is recorded currently (defaults to "F").}

\item{male.string}{How sex for males is recorded currently (defaults to "M").}
}
\description{
Convert Sex Data to format used by our Functions
}
\examples{
data(pupfish) ## In geomorph.
ConvertSexFormat(pupfish$Sex)
}
