% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DShape.R
\name{DShape}
\alias{DShape}
\title{Effect Size of Sexual Shape Dimorphism}
\usage{
DShape(Coords, Sex)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
Roughly based on Cohen's d, a common effect size statistic, another way of representing sexual shape dimorphism that is more readily comparable between studies.
}
