% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Norm}
\usage{
euclidean(vector)
}
\arguments{
\item{vector}{numerical vector of numbers}
}
\description{
The way you calculate distances in R^n, aka the square root of the sum of squares.
}
\examples{
euclidean(c(1,2,3.2))
}
