% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglefuncs.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Find the hypotenuse of a right-angled triangle in Euclidean space from the
other two sides.}
\usage{
hypotenuse(opposite, adjacent)
}
\arguments{
\item{opposite}{A number.}

\item{adjacent}{A number.}
}
\value{
The hypotenuse of the right-angled triangle with sides
\code{opposite} and \code{adjacent}.
}
\description{
Find the hypotenuse of a right-angled triangle in Euclidean space from the
other two sides.
}
\examples{
hypotenuse(1, 1)
hypotenuse(10, 1)
}
