% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglefuncs.R
\name{phi}
\alias{phi}
\title{Find the angle between opposite and hypotenuse of a right-angled triangle in
Euclidean space.}
\usage{
phi(opposite, adjacent)
}
\arguments{
\item{opposite}{A number.}

\item{adjacent}{A number.}
}
\value{
The angle between side \code{opposite} and the hypotenuse.
}
\description{
Find the angle between opposite and hypotenuse of a right-angled triangle in
Euclidean space.
}
\examples{
phi(1, 1)
phi(10, 1)
}
