% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSDSize.R
\name{PSDSize}
\alias{PSDSize}
\title{Pooled Standard Deviation for Size with respect to Sex}
\usage{
PSDSize(Size, Sex)
}
\arguments{
\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
The standard deviation of size after removing the difference between the sexes and normalizing the individual size to have a mean of 1.
}
