% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDLM.R
\name{SSDLM}
\alias{SSDLM}
\title{Sexual Size Dimorphism Calculator}
\usage{
SSDLM(Size, Sex)
}
\arguments{
\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{character or factor vector of individual's sexes given as 'f' for female and 'm' for male}
}
\description{
Calculates the SDI measure of sexual size dimorphism (Lovich and Gibbons 1992).  In order to calculate the log ratio instead use SSDLog.  Compared to SSDFunc, this is less susceptible to unbalanced designs.
}
\examples{
SSDFunc(Size, Sex)
}
