% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDLog.R
\name{SSDLog}
\alias{SSDLog}
\title{Sexual Size Dimorphism Calculator}
\usage{
SSDLog(SizeVec, SexVec, Base = exp(1), positiveF = F)
}
\arguments{
\item{SizeVec}{some numerical size measure given as a scalar quantity}

\item{SexVec}{sex values given either as a vector of characters or factors, must be formatted as 'f' for female or 'm' for male}

\item{Base}{the base for the logarithm (defaults to e = 2.718...)}

\item{positiveF}{Smith 1999 has the log ratio with male biased being positive and female biased being negative (suited to mammalian research), can reverse that by setting this to TRUE}
}
\description{
Calculates the log ratio of sexual dimorphism, log(M/F) (Smith 1999).  For the SDI (Lovich and Gibbons 1992), use SSDFunc().  Defaults to e as the base for the logarithm.
}
\examples{
SSD(Size, Sex)
}
