% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShD3DFunc.R
\name{SShD3DFunc}
\alias{SShD3DFunc}
\title{Calculating Sexual Shape Dimorphism}
\usage{
SShD3DFunc(Coords, Sex, Zeroed = TRUE)
}
\arguments{
\item{Coords}{Three-dimensional array of coordinates from geometric morphometric analysis (see gpagen() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}
}
\description{
This function directly calculates SShD.
}
\examples{
SShD3DFunc(Coords, Sex)
SShD3DFunc(Coords, Sex, Zeroed = TRUE)
SShD3DFunc(Coords, Sex, TRUE)
}
