% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDFast.R
\name{SShDFast}
\alias{SShDFast}
\title{Calculating Sexual Shape Dimorphism Fast}
\usage{
SShDFast(Coords, Sex)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
This function directly calculates SShD in the fastest manner (accuracy lost).
}
\examples{
SShDFast(Coords, Sex)
}
