% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDFunc.R
\name{SShDFunc}
\alias{SShDFunc}
\title{Calculating Sexual Shape Dimorphism}
\usage{
SShDFunc(Coords, SexVec, Zeroed = TRUE)
}
\arguments{
\item{Coords}{two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{SexVec}{character or factor vector of individual's sexes given as 'f' for female and 'm' for male}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}
}
\description{
This function directly calculates SShD.
}
\examples{
SShDFunc(Coords, Sex, Zeroed = TRUE)
SShDFunc(Coords, Sex, TRUE)
}
