% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDLM.R
\name{SShDLM}
\alias{SShDLM}
\title{Calculate SShD using the lm() function}
\usage{
SShDLM(Coords, Sex, Zeroed = T)
}
\arguments{
\item{Coords}{two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}

\item{SexVec}{vector of factors for male and female (can only be the two factors)}
}
\description{
All of this program's built in SShD functions that perform bootstrap analysis on the data use the lm() function for some of its nice properties.  This function allows you to reproduce the results of the lm() function-based SShD calculation for a single sample.
}
