% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDwSE.R
\name{SShDwSE}
\alias{SShDwSE}
\title{Calculate the Standard Error of the SShD (and SSD)}
\usage{
SShDwSE(Coords, Sex, Size = NULL, rep = 1000, ...)
}
\arguments{
\item{Coords}{Coordinates formatted according to the function two.d.array() from geomorph.}

\item{rep}{The number of times that you wish the resampling to be done, defaults to 1000 but more is better except for very small samples.}

\item{SexVec}{Sex of each individual given in standardized form (see ConvertSexFormat() from this package).}

\item{SizeVec}{Optional size argument, a numerical vector for the size of each individual, if you want the standard error of the size.}
}
\description{
The SShD or SSD is worth very little without a measure of error, this function runs a bootstrap in order to get a value for their standard errors.
}
\examples{
SShDwSE(Coords, Sex, Size, rep = 10000)
}
