% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratDShape.R
\name{StratDShape}
\alias{StratDShape}
\title{Effect Size of Sexual Shape Dimorphism by Stratum}
\usage{
StratDShape(Strata, Coords, Sex)
}
\arguments{
\item{Strata}{}

\item{Coords}{}

\item{Sex}{}
}
\description{
Roughly based on Cohen's d, a common effect size statistic, another way of representing sexual shape dimorphism that is more readily comparable between studies, calculated by group.
}
