% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratPSDShape.R
\name{StratPSDShape}
\alias{StratPSDShape}
\title{Calculates the Pooled Standard Deviation of Shape by Group}
\usage{
StratPSDShape(Strata, Coords, Sex)
}
\arguments{
\item{Strata}{A character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SShD and SSD}

\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
Calculates the Pooled Standard Deviation of Shape by Group
}
