% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratPSDSize.R
\name{StratPSDSize}
\alias{StratPSDSize}
\title{Calculate the Pooled Standard Deviation for Size for each Group}
\usage{
StratPSDSize(Strata, Size, Sex)
}
\arguments{
\item{Strata}{A character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SShD and SSD}

\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
Calculate the Pooled Standard Deviation for Size for each Group
}
