% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratSSD.R
\name{StratSSD}
\alias{StratSSD}
\title{Within-Group SSD Calculation}
\usage{
StratSSD(Strata, Size, Sex, Log = FALSE, ...)
}
\arguments{
\item{Strata}{A character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SSD}

\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Log}{Defaults to FALSE, in which case you use the Lovich and Gibbons SDI to calculate SSD, if changed to TRUE you use the log ratio.}

\item{...}{If Log is TRUE, then you may pass arguments to the SSDLog function this way.}
}
\description{
Within-Group SSD Calculation
}
\examples{
StratSSD(Strata, Size, Sex)
StratSSD(Strata, Size, Sex, Log = TRUE)
StratSSD(Strata, Size, Sex, Log = TRUE, positiveF = TRUE, base = 2)
}
