% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratSSDBoot.R
\name{StratSSDBoot}
\alias{StratSSDBoot}
\title{Stratified Standard Error calculator for SSD}
\usage{
StratSSDBoot(Size, Sex, Strata, rep = 1000)
}
\arguments{
\item{Size}{a numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{a character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Strata}{a character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SSD}

\item{rep}{number of times to perform the resampling procedure}
}
\description{
Stratified Standard Error calculator for SSD
}
