% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratSShDBoot.R
\name{StratSShDBoot}
\alias{StratSShDBoot}
\title{Standard Error of SShD}
\usage{
StratSShDBoot(Coords, Sex, NonSexFactor, rep = 1000)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{NonSexFactor}{}

\item{rep}{The number of times that you wish the resampling to be done, defaults to 1000 but more is better except for very small samples.}
}
\description{
Standard Error of SShD
}
