% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratZStat.R
\name{StratZStat}
\alias{StratZStat}
\title{Calculate Geomorph's Z-Statistic for Sexual Shape Dimorphism within Groups}
\usage{
StratZStat(Strata, Coords, Sex)
}
\arguments{
\item{Strata}{}

\item{Coords}{}

\item{Sex}{}
}
\description{
Calculate Geomorph's Z-Statistic for Sexual Shape Dimorphism within Groups
}
