% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglefuncs.R
\name{theta}
\alias{theta}
\title{Find the angle between adjacent and hypotenuse of a right-angled triangle in
Euclidean space.}
\usage{
theta(opposite, adjacent)
}
\arguments{
\item{opposite}{A number.}

\item{adjacent}{A number.}
}
\value{
The angle between side \code{adjacent} and the hypotenuse.
}
\description{
Find the angle between adjacent and hypotenuse of a right-angled triangle in
Euclidean space.
}
\examples{
theta(1, 1)
theta(10, 1)
}
