% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSShD.R
\name{zeroSShD}
\alias{zeroSShD}
\title{Calculates Shape Diversity regardless of Sex}
\usage{
zeroSShD(Coords, SexVec)
}
\arguments{
\item{Coords}{two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{SexVec}{character vector of individual's sexes given as 'f' for female and 'm' for male}
}
\description{
Determines the shape diversity in the population in a way that can be used to provide a benchmark for how much sexual shape dimorphism there is over and above the overall shape diversity.
}
\examples{
zeroSShD(Coords, SexVec)
}
