% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZStat.R
\name{ZStat}
\alias{ZStat}
\title{Geomorph's Z-statistic for Sexual Shape Dimorphism}
\usage{
ZStat(Coords, Sex)
}
\arguments{
\item{Coords}{}

\item{Sex}{}
}
\description{
Geomorph's Z-statistic for Sexual Shape Dimorphism
}
