% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDFunc.R
\name{SSDFunc}
\alias{SSDFunc}
\title{Sexual Size Dimorphism Calculator}
\usage{
SSDFunc(SizeVec, SexVec)
}
\arguments{
\item{SizeVec}{some numerical size measure given as a scalar quantity}

\item{SexVec}{sex values given either as a vector of characters or factors, must be formatted as 'f' for female or 'm' for male}
}
\description{
Calculates the SDI measure of sexual size dimorphism (Lovich and Gibbons 1992).  In order to calculate the log ratio instead use SSDLog.
}
\examples{
SSD(Size, Sex)
}
