% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDLog.R
\name{SSDLog}
\alias{SSDLog}
\title{Sexual Size Dimorphism Calculator}
\usage{
SSDLog(Size, Sex, Base = exp(1), positiveF = F)
}
\arguments{
\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Base}{the base for the logarithm (defaults to e = 2.718...)}

\item{positiveM}{Smith 1999 has the log ratio with male biased being positive and female biased being negative (suited to mammalian research), can have the statistic operate in that manner by setting this to TRUE}
}
\description{
Calculates the log ratio of sexual dimorphism, log(F/M) (adapted with alterations from Smith 1999).  For the SDI (Lovich and Gibbons 1992), use SSDFunc().  Defaults to e as the base for the logarithm.
}
\examples{
SSDLog(Size, Sex)
}
