% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShD3DFunc.R
\name{SShD3DFunc}
\alias{SShD3DFunc}
\title{Calculating Sexual Shape Dimorphism}
\usage{
SShD3DFunc(Coords, SexVec, Zeroed = TRUE)
}
\arguments{
\item{Coords}{three-dimensional array of coordinates from geometric morphometric analysis (see gpagen() in geomorph for correct formatting)}

\item{SexVec}{character vector of individual's sexes given as 'f' for female and 'm' for male}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}
}
\description{
This function directly calculates SShD.
}
\examples{
SShD3DFunc(Coords, Sex)
SShD3DFunc(Coords, Sex, Zeroed = TRUE)
SShD3DFunc(Coords, Sex, TRUE)
}
