% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDFast.R
\name{SShDFast}
\alias{SShDFast}
\title{Calculating Sexual Shape Dimorphism Fast}
\usage{
SShDFast(Coords, SexVec)
}
\arguments{
\item{Coords}{two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{SexVec}{character vector of individual's sexes given as 'f' for female and 'm' for male}
}
\description{
This function directly calculates SShD in the fastest manner (accuracy lost).
}
\examples{
SShDFast(Coords, Sex)
}
