% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDFunc.R
\name{SShDFunc}
\alias{SShDFunc}
\title{Calculating Sexual Shape Dimorphism}
\usage{
SShDFunc(Coords, Sex, Zeroed = TRUE)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}
}
\description{
This function directly calculates SShD.  However it may be a poor estimator for unbalanced designs where there are more of one sex than the other, SShDLM() is more robust.
}
\examples{
SShDFunc(Coords, Sex, Zeroed = TRUE)
SShDFunc(Coords, Sex, TRUE)
}
