% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShDwSE.R
\name{SShDwSE}
\alias{SShDwSE}
\title{Calculate the Standard Error of the SShD (and SSD)}
\usage{
SShDwSE(Coords, Sex, Size = NULL, rep = 1000, ...)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Size}{Optional size argument, a numerical vector giving the size of each individual as a scalar quantity}

\item{rep}{The number of times that you wish the resampling to be done, defaults to 1000 but more is better except for very small samples.}
}
\description{
The SShD or SSD is worth very little without a measure of error, this function runs a bootstrap in order to get a value for their standard errors.
}
\examples{
SShDwSE(Coords, Sex, Size, rep = 10000, Zeroed = FALSE)
}
