% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratSShD.R
\name{StratSShD}
\alias{StratSShD}
\title{Within-Group SShD calculation}
\usage{
StratSShD(Strata, Coords, Sex, Zeroed = TRUE)
}
\arguments{
\item{Strata}{A character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SShD}

\item{Coords}{two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{character or factor vector of individual's sexes given as 'f' for female and 'm' for male}

\item{Zeroed}{a logical value stating whether you wish to correct the SShD for the variation among individuals regardless of sex (defaults to TRUE)}
}
\description{
Within-Group SShD calculation
}
\examples{
StratSShD(Strata, Coords, Sex, T)
StratSShD(Strata, Coords, Sex, Zeroed = FALSE)
}
