% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratSShDwSE.R
\name{StratSShDwSE}
\alias{StratSShDwSE}
\title{Calculate stratified Standard Errors for SShD and SSD}
\usage{
StratSShDwSE(Coords, Sex, Size, Strata, rep = 1000, print.progress = T,
  logSSD = F, ...)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}

\item{Size}{A numerical vector giving the size of each individual as a scalar quantity}

\item{Strata}{A character or factor vector giving the factor by which we wish to divide individuals into groups before calculating SShD and SSD}

\item{rep}{Number of resamplings you wish to do}
}
\description{
Calculate stratified Standard Errors for SShD and SSD
}
