% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero3DSShD.R
\name{zero3DSShD}
\alias{zero3DSShD}
\title{Calculates Shape Diversity regardless of Sex}
\usage{
zero3DSShD(Coords, SexVec)
}
\arguments{
\item{Coords}{three-dimensional array of coordinates from geometric morphometric analysis (see gpagen() in geomorph for correct formatting)}

\item{SexVec}{character vector of individual's sexes given as 'f' for female and 'm' for male}
}
\description{
Determines the shape diversity in the population in a way that can be used to provide a benchmark for how much sexual shape dimorphism there is over and above the overall shape diversity.
}
\examples{
zero3DSShD(Coords, SexVec)
}
