% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSShDLM.R
\name{zeroSShDLM}
\alias{zeroSShDLM}
\title{Calculates Shape Diversity regardless of Sex}
\usage{
zeroSShDLM(Coords, Sex)
}
\arguments{
\item{Coords}{Two-dimensional array of coordinates from geometric morphometric analysis (see two.d.array() in geomorph for correct formatting)}

\item{Sex}{A character or factor vector recording sex for each individual as 'm' or 'f'}
}
\description{
Determines the shape diversity in the population in a way that can be used to provide a benchmark for how much sexual shape dimorphism there is over and above the overall shape diversity.
}
\examples{
zeroSShDLM(Coords, Sex)
}
